+++ src/Request/Internal.php
@@ -391,9 +391,10 @@ class Internal extends RequestCollection
                         ->addPost('attached_media', json_encode($attachedMedia))
                         ->addPost('story_sticker_ids', 'media_simple_'.reset($attachedMedia)['media_id']);
                 }
-                if (isset($externalMetadata['share_to_fb_destination_id']) && isset($externalMetadata['fb_access_token'])) {
+                if (isset($externalMetadata['share_to_fb_destination_id'])) {
                     $request->addPost('share_to_fb_destination_id', $externalMetadata['share_to_fb_destination_id'])
-                            ->addPost('fb_access_token', $externalMetadata['fb_access_token'])
+                            ->addPost('xpost_surface', 'auto_xpost')
+                            ->addPost('share_to_facebook', 1)
                             ->addPost('share_to_fb_destination_type', 'USER');
                 }
                 break;
@@ -899,9 +900,10 @@ class Internal extends RequestCollection
                         ->addPost('attached_media', json_encode($attachedMedia))
                         ->addPost('story_sticker_ids', 'media_simple_'.reset($attachedMedia)['media_id']);
                 }
-                if (isset($externalMetadata['share_to_fb_destination_id']) && isset($externalMetadata['fb_access_token'])) {
+                if (isset($externalMetadata['share_to_fb_destination_id'])) {
                     $request->addPost('share_to_fb_destination_id', $externalMetadata['share_to_fb_destination_id'])
-                            ->addPost('fb_access_token', $externalMetadata['fb_access_token'])
+                            ->addPost('xpost_surface', 'auto_xpost')
+                            ->addPost('share_to_facebook', 1)
                             ->addPost('share_to_fb_destination_type', 'USER');
                 }
                 break;